Remove old implementation
