+++ src/Moq.Analyzers/AsShouldBeUsedOnlyForInterfaceAnalyzer.cs
@@ -1,3 +1,5 @@
+﻿using Microsoft.CodeAnalysis.Operations;
+
 namespace Moq.Analyzers;
 
 /// <summary>
@@ -9,8 +11,6 @@ public class AsShouldBeUsedOnlyForInterfaceAnalyzer : DiagnosticAnalyzer
     internal const string RuleId = "Moq1300";
     private const string Title = "Moq: Invalid As type parameter";
     private const string Message = "Mock.As() should take interfaces only";
-
-    private static readonly MoqMethodDescriptorBase MoqAsMethodDescriptor = new MoqAsMethodDescriptor();
 
     private static readonly DiagnosticDescriptor Rule = new(
         RuleId,
@@ -29,43 +29,58 @@ public class AsShouldBeUsedOnlyForInterfaceAnalyzer : DiagnosticAnalyzer
     {
         context.EnableConcurrentExecution();
         context.ConfigureGeneratedCodeAnalysis(GeneratedCodeAnalysisFlags.None);
-        context.RegisterSyntaxNodeAction(Analyze, SyntaxKind.InvocationExpression);
+
+        context.RegisterCompilationStartAction(static context =>
+        {
+            // Ensure Moq is referenced in the compilation
+            ImmutableArray<INamedTypeSymbol> mockTypes = context.Compilation.GetMoqMock();
+            if (mockTypes.IsEmpty)
+            {
+                return;
+            }
+
+            // Look for the Mock.As() method and provide it to Analyze to avoid looking it up multiple times.
+            ImmutableArray<IMethodSymbol> asMethods = mockTypes
+                .SelectMany(mockType => mockType.GetMembers("As"))
+                .OfType<IMethodSymbol>()
+                .Where(method => method.IsGenericMethod)
+                .ToImmutableArray();
+            if (asMethods.IsEmpty)
+            {
+                return;
+            }
+
+            context.RegisterOperationAction(context => Analyze(context, asMethods), OperationKind.Invocation);
+        });
     }
 
-    [System.Diagnostics.CodeAnalysis.SuppressMessage("Maintainability", "AV1500:Member or local function contains too many statements", Justification = "Tracked in https://github.com/rjmurillo/moq.analyzers/issues/90")]
-    private static void Analyze(SyntaxNodeAnalysisContext context)
+    private static void Analyze(OperationAnalysisContext context, ImmutableArray<IMethodSymbol> wellKnownAsMethods)
     {
-        if (context.Node is not InvocationExpressionSyntax invocationExpression)
+        if (context.Operation is not IInvocationOperation invocationOperation)
         {
             return;
         }
 
-        if (invocationExpression.Expression is not MemberAccessExpressionSyntax memberAccessSyntax)
+        IMethodSymbol targetMethod = invocationOperation.TargetMethod;
+        if (!targetMethod.IsInstanceOf(wellKnownAsMethods))
         {
             return;
         }
 
-        if (!MoqAsMethodDescriptor.IsMatch(context.SemanticModel, memberAccessSyntax, context.CancellationToken))
+        ImmutableArray<ITypeSymbol> typeArguments = targetMethod.TypeArguments;
+        if (typeArguments.Length != 1)
         {
             return;
         }
 
-        if (!memberAccessSyntax.Name.TryGetGenericArguments(out SeparatedSyntaxList<TypeSyntax> typeArguments))
+        if (typeArguments[0] is ITypeSymbol { TypeKind: not TypeKind.Interface })
         {
-            return;
-        }
+            // Try to locate the type argument in the syntax tree to report the diagnostic at the correct location.
+            // If that fails for any reason, report the diagnostic on the operation itself.
+            NameSyntax? memberName = context.Operation.Syntax.DescendantNodes().OfType<MemberAccessExpressionSyntax>().Select(mae => mae.Name).FirstOrDefault();
+            Location location = memberName?.GetLocation() ?? invocationOperation.Syntax.GetLocation();
 
-        if (typeArguments.Count != 1)
-        {
-            return;
-        }
-
-        TypeSyntax typeArgument = typeArguments[0];
-        SymbolInfo symbolInfo = context.SemanticModel.GetSymbolInfo(typeArgument, context.CancellationToken);
-
-        if (symbolInfo.Symbol is ITypeSymbol { TypeKind: not TypeKind.Interface })
-        {
-            context.ReportDiagnostic(Diagnostic.Create(Rule, typeArgument.GetLocation()));
+            context.ReportDiagnostic(Diagnostic.Create(Rule, location));
         }
     }
 }
