+++ tests/Moq.Analyzers.Benchmarks/Moq1300Benchmarks.cs
@@ -13,9 +13,7 @@ public class Moq1300Benchmarks
 {
     private static CompilationWithAnalyzers? BaselineCompilation { get; set; }
 
-    private static CompilationWithAnalyzers? OldTestCompilation { get; set; }
-
-    private static CompilationWithAnalyzers? NewTestCompilation { get; set; }
+    private static CompilationWithAnalyzers? TestCompilation { get; set; }
 
     [IterationSetup]
     [SuppressMessage("Usage", "VSTHRD002:Avoid problematic synchronous waits", Justification = "Async setup not supported in BenchmarkDotNet.See https://github.com/dotnet/BenchmarkDotNet/issues/2442.")]
@@ -46,40 +44,18 @@ internal class {name}
 "));
         }
 
-        (BaselineCompilation, OldTestCompilation) =
+        (BaselineCompilation, TestCompilation) =
             BenchmarkCSharpCompilationCreator<AsShouldBeUsedOnlyForInterfaceAnalyzer>
             .CreateAsync(sources.ToArray())
             .GetAwaiter()
             .GetResult();
-
-        (_, NewTestCompilation) =
-            BenchmarkCSharpCompilationCreator<AsShouldBeUsedOnlyForInterfaceAnalyzer2>
-            .CreateAsync(sources.ToArray())
-            .GetAwaiter()
-            .GetResult();
     }
 
     [Benchmark]
-    public async Task OldMoq1300WithDiagnostics()
+    public async Task Moq1300WithDiagnostics()
     {
         ImmutableArray<Diagnostic> diagnostics =
-            (await OldTestCompilation!
-            .GetAnalysisResultAsync(CancellationToken.None)
-            .ConfigureAwait(false))
-            .AssertValidAnalysisResult()
-            .GetAllDiagnostics();
-
-        if (diagnostics.Length != Constants.NumberOfCodeFiles)
-        {
-            throw new InvalidOperationException($"Expected '{Constants.NumberOfCodeFiles:N0}' analyzer diagnostics but found '{diagnostics.Length}'");
-        }
-    }
-
-    [Benchmark]
-    public async Task NewMoq1300WithDiagnostics()
-    {
-        ImmutableArray<Diagnostic> diagnostics =
-            (await NewTestCompilation!
+            (await TestCompilation!
             .GetAnalysisResultAsync(CancellationToken.None)
             .ConfigureAwait(false))
             .AssertValidAnalysisResult()
